cerebrum-pi/requirements.txt
# Cerebrum CM4 Requirements
# Role: Lightweight orchestrator + API server
# Heavy inference: Handled by VPS backend (127.0.0.1:9000)

# Core dependencies
# torch>=2.1.0
psutil>=5.9.0
pydantic>=2.0.0

# Model inference
# llama-cpp-python --only-binary=llama-cpp-python

# API layer
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Symbolic reasoning
z3-solver>=4.12.0
